+++ test/11-functions.js
@@ -2,6 +2,13 @@
 
 const {expect} = require("chai");
 
+let asyncFunc = false;
+let asyncGen  = false;
+try{
+	asyncFunc = Function("return async function bar(){ return 20; };")();
+	asyncGen  = Function("return async function* bar(){ yield 20; };")();
+} catch(e){}
+
 describe("Functions", () => {
 	it("prints named functions", () => {
 		const func = function(i){ return i + 2; };
@@ -31,16 +38,6 @@ describe("Functions", () => {
 		}`);
 	});
 	
-	it("recognises generator functions", () => {
-		const obj = { money: function* generator(){ yield $20; } };
-		expect(obj).to.print(`{
-			money: function*(){
-				length: 0
-				name: "generator"
-			}
-		}`);
-	});
-	
 	it("avoids breaking on absent constructors", () => {
 		const obj = {get constructor(){ return undefined; }};
 		expect(obj).to.print(`{
@@ -48,35 +45,43 @@ describe("Functions", () => {
 		}`);
 	});
 
-
-	describe("Classes", function(){
-		class Example{
-			constructor(){
-				this.name = "Foo";
+	it("identifies generator functions", () => {
+		const foo = function* bar(){ yield 20; };
+		expect(foo).to.print(`function*(){
+			length: 0
+			name: "bar"
+		}`);
+		expect({method: foo}).to.print(`{
+			method: function*(){
+				length: 0
+				name: "bar"
 			}
-		}
-		
-		class ExtendedExample extends Example{
-			constructor(){
-				super();
-				this.name = "Bar";
-			}
-		}
-
-		it("prints class instances", () => {
-			expect(new Example()).to.print(`
-				Example{
-					name: "Foo"
-				}
-			`);
-		});
-		
-		it("prints instances of subclasses", () => {
-			expect(new ExtendedExample()).to.print(`
-				ExtendedExample{
-					name: "Bar"
-				}
-			`);
-		});
+		}`);
 	});
+	
+	it("identifies asynchronous functions", asyncFunc ? () => {
+		expect(asyncFunc).to.print(`async function(){
+			length: 0
+			name: "bar"
+		}`);
+		expect({method: asyncFunc}).to.print(`{
+			method: async function(){
+				length: 0
+				name: "bar"
+			}
+		}`);
+	} : undefined);
+	
+	it("identifies asynchronous generators", asyncGen ? () => {
+		expect(asyncGen).to.print(`async function*(){
+			length: 0
+			name: "bar"
+		}`);
+		expect({method: asyncGen}).to.print(`{
+			method: async function*(){
+				length: 0
+				name: "bar"
+			}
+		}`);
+	} : undefined);
 });